Move some helper functions for dealing with members to separate util
